MPC_Allocator: Patch from Julien to port allocator code on windows

+++ mpc/MPC_Allocator/src/sctk_alloc_common.h
@@ -17,6 +17,7 @@
 /* #                                                                      # */
 /* # Authors:                                                             # */
 /* #   - Valat Sébastien sebastien.valat@cea.fr                           # */
+/* #   - Adam Julien julien.adam.ocre@cea.fr                              # */
 /* #                                                                      # */
 /* ######################################################################## */
 
@@ -40,7 +41,7 @@ extern "C"
 **/
 #define SCTK_ALLOC_SMALL_CHUNK_SIZE 32
 /** Define the size of a macro bloc (2Mo by default) **/
-#define SCTK_MACRO_BLOC_SIZE (2*1024*1024UL)
+#define SCTK_MACRO_BLOC_SIZE (2*1024ULL*1024ULL)
 /** Number of size class for the free list. **/
 #define SCTK_ALLOC_NB_FREE_LIST 48
 /** Minimal size of blocks. **/
@@ -50,9 +51,9 @@ extern "C"
 /** Define the considered page size. **/
 #define SCTK_ALLOC_PAGE_SIZE 4096
 /** Size allocated for region headers. **/
-#define SCTK_REGION_HEADER_SIZE (4*1024*1024ul)
+#define SCTK_REGION_HEADER_SIZE (4*1024ULL*1024ULL)
 /** Size of a region header (1TB for 2MB macro-blocs and 1 pointer per entry) **/
-#define SCTK_REGION_SIZE (1024UL*1024UL*1024UL*1024UL)
+#define SCTK_REGION_SIZE (1024ULL*1024ULL*1024ULL*1024ULL)
 /** Number of entries of a region header. **/
 #define SCTK_REGION_HEADER_ENTRIES ((SCTK_REGION_SIZE) / SCTK_MACRO_BLOC_SIZE)
 /** Base addresse for the current process heap based on mmap. **/
@@ -95,9 +96,9 @@ extern "C"
 
 /************************** TYPES **************************/
 /** Type for size member, must be 64bit type to maintain alignment coherency. **/
-typedef unsigned long sctk_size_t;
+typedef size_t sctk_size_t;
 /** Type for address member, must be 64bit type to maintain alignement coherency. **/
-typedef unsigned long sctk_addr_t;
+typedef size_t sctk_addr_t;
 /** Type for short size member, must be 8bits type to maintain alignement coherency. **/
 typedef unsigned char sctk_short_size_t;
